% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_jk.R
\name{compute_Jk}
\alias{compute_Jk}
\title{compute_Jk}
\usage{
compute_Jk(
  v,
  s_k,
  p_k,
  Y,
  alpha,
  w,
  m,
  use0,
  use1,
  domain,
  select_domain,
  diss_d0_d1_L2,
  c_k = NULL,
  keep_k = NULL
)
}
\arguments{
\item{v}{list of two elements: v0=v(x), v1=v'(x), matrices with d columns.}

\item{s_k}{integer vector: shift vector for motif k.}

\item{p_k}{numeric vector: membership vector for motif k.}

\item{Y}{list of N lists of two elements: Y0=y_i(x), Y1=y'_i(x), matrices with d columns, for d-dimensional curves.}

\item{alpha}{double: weight coefficient between d0_L2 and d1_L2.}

\item{w}{numeric vector: weights for the dissimilarity index in the different dimensions (w>0).}

\item{c_k}{integer vector: minimum length of supp(y_shifted) and supp(v) intersection.}

\item{keep_k}{boolean vector: check c_k only when keep=TRUE for y_shifted.}

\item{m>1}{integer vector: weighting exponent in least-squares functional.}
}
\value{
The value of the objective function J
}
\description{
Compute the objective function J for the motif k.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
