% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_motif.R
\name{compute_motif}
\alias{compute_motif}
\title{compute_motif}
\usage{
compute_motif(v_dom, s_k, p_k, Y, m, use0, use1)
}
\arguments{
\item{v_dom}{TRUE for x in supp(v).}

\item{s_k}{shift vector for motif k.}

\item{p_k}{membership vector for motif k.}

\item{Y}{list of N lists of two elements, Y0=y_i(x), Y1=y'_i(x), matrices with d columns, for d-dimensional curves.}
}
\value{
The new motif v_new
}
\description{
Compute the new motif v_new.
}
\author{
Marzia Angela Cremona  & Francesca Chiaromonte
}
