% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_candidate_motifs.R
\name{find_candidate_motifs}
\alias{find_candidate_motifs}
\title{find_candidate_motifs}
\usage{
find_candidate_motifs(
  Y0,
  Y1 = NULL,
  K,
  c,
  n_init = 10,
  name = "results",
  names_var = "",
  probKMA_options = NULL,
  silhouette_align = FALSE,
  plot = TRUE,
  worker_number = NULL
)
}
\arguments{
\item{Y0}{list of N vectors, for univariate curves y_i(x), or list of N matrices with d columns,
for d-dimensional curves y_i(x),  with the evaluation of curves (all curves should be evaluated
on a uniform grid). When y_j(x)=NA in the dimension j, then y_j(x)=NA in ALL dimensions}

\item{Y1}{list of N vectors, for univariate derivative curves y'_i(x), or
list of N matrices with d columns, for d-dimensional derivatibe curves y'_i(x),
with the evaluation of the curves derivatives (all curves should be evaluated on a uniform grid).
When y'_j(x)=NA in the dimension j, then y'_j(x)=NA in ALL dimensions.
Must be provided when diss='d1_L2' or diss='d0_d1_L2'.}

\item{K}{vector with numbers of motifs that must be tested.}

\item{c}{vector with minimum motifs lengths that must be tested.}

\item{n_init}{number of random initialization for each combination of K and c.}

\item{name}{name of the folders when the results are saved.}

\item{names_var}{vector of length d, with names of the variables in the different dimensions.}

\item{probKMA_options}{list with options for probKMA (see the help of probKMA).}

\item{silhouette_align}{True or False. If True, try all possible alignments between the curve pieces
when calculating the adapted silhouette index on the results of probKMA}

\item{plot}{if TRUE, summary plots are drawn.}

\item{worker_number}{number of CPU cores to be used for parallelization (default number of CPU cores -1).
If worker_number=1, the function is run sequentially.}
}
\value{
A list containing: K, c, n_init and name;...

\item{times}{ list of execution times of ProbKMA for each combination of K, c, and n_init}

\item{silhouette_average_sd}{ list of the mean (silhouette_average) and standard deviation (silhouette_sd) of the silhouette indices for each execution of the ProbKMA function}
}
\description{
Run multiple times probKMA function with different K,c and initializations,
with the aim to find a set of candidate motifs.
If the folder name_KK_cc is already present and n result files are already present,
load them and continue with the n_init-n runs.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
