% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_diss_aligned.R
\name{find_diss_aligned}
\alias{find_diss_aligned}
\title{find_diss_aligned}
\usage{
find_diss_aligned(y, v, alpha, w, aligned, d, use0, use1)
}
\arguments{
\item{y}{list of two elements y0=y(x), y1=y'(x), matrices with d columns.}

\item{v}{list of two elements v0=v(x), v1=v'(x), matrices with d columns.}

\item{alpha}{weight coefficient between d0.L2 and d1.L2.}

\item{w}{weights for the dissimilarity index in the different dimensions (w>0).}

\item{aligned}{if TRUE, curves are already aligned. Else, the shortest curve is aligned inside the longest.}
}
\value{
shift and dissimilarity
}
\description{
Find dissimilarity between multidimensional curves (dimension=d), without alignment unless
their lengths are different. To be used by probKMA_silhouette fucntion.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
