% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_shift_warp_min.R
\name{find_shift_warp_min}
\alias{find_shift_warp_min}
\title{find_shift_warp_min}
\usage{
find_shift_warp_min(
  y,
  v,
  w,
  c_k,
  K,
  d,
  max_gap,
  alpha,
  use0,
  use1,
  domain,
  select_domain,
  diss_d0_d1_L2
)
}
\arguments{
\item{y}{list of two elements y0=y(x), y1=y'(x), matrices with d columns.}

\item{v}{list of two elements v0=v(x), v1=v'(x), matrices with d columns.}

\item{w}{weights for the dissimilarity index in the different dimensions (w>0).}

\item{c_k}{minimum length of supp(y_shifted) and supp(v) intersection.}

\item{alpha}{weight coefficient between d0.L2 and d1.L2.}
}
\value{
Shift warping and dissimilarity
}
\description{
Find shift warping minimizing dissimilarity between multidimensional curves (dimension=d).
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
