\name{ProbKMA.package-package}
\alias{ProbKMA.package-package}
\alias{ProbKMA.package}
\docType{package}
\title{\packageTitle{ProbKMA.package}}
\description{\packageDescription{ProbKMA.package}}
\section{Package Content}{\packageIndices{ProbKMA.package}}
\author{\packageAuthor{ProbKMA.package}}
\section{Maintainer}{\packageMaintainer{ProbKMA.package}}
\keyword{package}
